+++ examples/resources/nxos_feature_lldp/import.sh
@@ -1,0 +1,1 @@
+terraform import nxos_feature_lldp.example "sys/fm/lldp"