+++ examples/resources/nxos_feature_lldp/resource.tf
@@ -1,0 +1,3 @@
+resource "nxos_feature_lldp" "example" {
+  admin_state = "enabled"
+}